+++ lib/generators/statesman/mongoid_transition_generator.rb
@@ -10,7 +10,7 @@ module Statesman
     argument :parent, type: :string, desc: "Your parent model name"
     argument :klass, type: :string, desc: "Your transition model name"
 
-    source_root File.expand_path("../templates", __FILE__)
+    source_root File.expand_path("templates", __dir__)
 
     def create_model_file
       template("mongoid_transition_model.rb.erb", model_file_name)