+++ lib/statesman/adapters/active_record_transition.rb
@@ -3,8 +3,15 @@ require "json"
 module Statesman
   module Adapters
     module ActiveRecordTransition
-      def self.included(base)
-        base.send(:serialize, :metadata, JSON)
+      DEFAULT_UPDATED_TIMESTAMP_COLUMN = :updated_at
+
+      extend ActiveSupport::Concern
+
+      included do
+        serialize :metadata, JSON
+
+        class_attribute :updated_timestamp_column
+        self.updated_timestamp_column = DEFAULT_UPDATED_TIMESTAMP_COLUMN
       end
     end
   end