+++ spec/generators/statesman/mongoid_transition_generator_spec.rb
@@ -7,7 +7,7 @@ describe Statesman::MongoidTransitionGenerator, type: :generator do
     before { run_generator %w[Yummy::Bacon Yummy::BaconTransition] }
     subject { file("app/models/yummy/bacon_transition.rb") }
 
-    it { is_expected.not_to contain(%r{:yummy/bacon}) }
+    it { is_expected.to_not contain(%r{:yummy/bacon}) }
     it { is_expected.to contain(/class_name: 'Yummy::Bacon'/) }
   end
 
@@ -15,7 +15,7 @@ describe Statesman::MongoidTransitionGenerator, type: :generator do
     before { run_generator %w[Bacon BaconTransition] }
     subject { file("app/models/bacon_transition.rb") }
 
-    it { is_expected.not_to contain(/class_name:/) }
-    it { is_expected.not_to contain(/CreateYummy::Bacon/) }
+    it { is_expected.to_not contain(/class_name:/) }
+    it { is_expected.to_not contain(/CreateYummy::Bacon/) }
   end
 end