+++ spec/statesman/adapters/active_record_queries_spec.rb
@@ -6,14 +6,9 @@ describe Statesman::Adapters::ActiveRecordQueries, active_record: true do
     prepare_transitions_table
     prepare_other_model_table
     prepare_other_transitions_table
-  end
 
-  before do
     Statesman.configure { storage_adapter(Statesman::Adapters::ActiveRecord) }
-  end
-  after { Statesman.configure { storage_adapter(Statesman::Adapters::Memory) } }
 
-  before do
     MyActiveRecordModel.send(:include, Statesman::Adapters::ActiveRecordQueries)
     MyActiveRecordModel.class_eval do
       def self.transition_class
@@ -36,9 +31,11 @@ describe Statesman::Adapters::ActiveRecordQueries, active_record: true do
         :initial
       end
     end
+
+    MyActiveRecordModel.send(:has_one, :other_active_record_model)
+    OtherActiveRecordModel.send(:belongs_to, :my_active_record_model)
   end
-  before { MyActiveRecordModel.send(:has_one, :other_active_record_model) }
-  before { OtherActiveRecordModel.send(:belongs_to, :my_active_record_model) }
+  after { Statesman.configure { storage_adapter(Statesman::Adapters::Memory) } }
 
   let!(:model) do
     model = MyActiveRecordModel.create
@@ -66,7 +63,7 @@ describe Statesman::Adapters::ActiveRecordQueries, active_record: true do
       subject { MyActiveRecordModel.in_state(:succeeded) }
 
       it { is_expected.to include model }
-      it { is_expected.not_to include other_model }
+      it { is_expected.to_not include other_model }
     end
 
     context "given multiple states" do
@@ -104,12 +101,14 @@ describe Statesman::Adapters::ActiveRecordQueries, active_record: true do
   describe ".not_in_state" do
     context "given a single state" do
       subject { MyActiveRecordModel.not_in_state(:failed) }
+
       it { is_expected.to include model }
-      it { is_expected.not_to include other_model }
+      it { is_expected.to_not include other_model }
     end
 
     context "given multiple states" do
       subject { MyActiveRecordModel.not_in_state(:succeeded, :failed) }
+
       it do
         is_expected.to match_array([initial_state_model,
                                     returned_to_initial_model])
@@ -118,6 +117,7 @@ describe Statesman::Adapters::ActiveRecordQueries, active_record: true do
 
     context "given an array of states" do
       subject { MyActiveRecordModel.not_in_state(%i[succeeded failed]) }
+
       it do
         is_expected.to match_array([initial_state_model,
                                     returned_to_initial_model])
@@ -143,6 +143,7 @@ describe Statesman::Adapters::ActiveRecordQueries, active_record: true do
 
     describe ".in_state" do
       subject(:query) { MyActiveRecordModel.in_state(:succeeded) }
+
       specify { expect { query }.to_not raise_error }
     end
   end